Chapter 6 - created login view login guard

+++ online-shop/src/app/components/product-details/product-details.component.ts
@@ -6,6 +6,7 @@ import { ShoppingCartService } from '../../services/shopping-cart.service';
 import { HttpClient } from '@angular/common/http';
 import { ErrorHandlerService } from 'src/app/services/error-handler.service';
 import { ProductService } from 'src/app/services/product.service';
+import { LoginService } from 'src/app/services/login.service';
 
 @Component({
   selector: 'app-product-details',
@@ -18,6 +19,7 @@ export class ProductDetailsComponent implements OnInit {
   deleted: boolean = false;
   errorMessage: string = '';
   product = {} as Product;
+  admin: boolean = false;
 
   constructor(
     private route: ActivatedRoute,
@@ -25,10 +27,13 @@ export class ProductDetailsComponent implements OnInit {
     private errorHandlerService: ErrorHandlerService,
     private router: Router,
     private http: HttpClient,
-    private productService: ProductService
+    private productService: ProductService,
+    private loginService: LoginService
   ) {}
 
   ngOnInit(): void {
+    this.admin = this.loginService.isAdmin();
+    console.log('Admin?', this.admin);
     this.id = this.route.snapshot.params.id;
     this.productService.getProduct(this.id).subscribe(
       (product: Product) => {
@@ -41,8 +46,12 @@ export class ProductDetailsComponent implements OnInit {
   }
 
   addToCart(product: Product): void {
-    this.shoppingCartService.addToCart(product);
-    window.alert('Your product has been added to the cart!');
+    if (this.loginService.isCustomer()) {
+      this.shoppingCartService.addToCart(product);
+      window.alert('Your product has been added to the cart!');
+    } else {
+      window.alert('Warning! Not authorized!');
+    }
   }
 
   edit(): void {
@@ -50,13 +59,17 @@ export class ProductDetailsComponent implements OnInit {
   }
 
   delete(): void {
-    this.productService.deleteProduct(this.id).subscribe(
-      () => {
-        this.deleted = true;
-      },
-      (error) => {
-        this.errorMessage = this.errorHandlerService.handleError(error);
-      }
-    );
+    if (this.loginService.isAdmin()) {
+      this.productService.deleteProduct(this.id).subscribe(
+        () => {
+          this.deleted = true;
+        },
+        (error) => {
+          this.errorMessage = this.errorHandlerService.handleError(error);
+        }
+      );
+    } else {
+      window.alert('Warning! Not authorized!');
+    }
   }
 }
